@@ -24,6 +24,8 @@ module.exports = {
   authorizeServiceAccount: resourceServer.authorizeServiceAccount,
   controlledAccountLogin: resourceServer.controlledAccountLogin,
   oidcLoginProviders: resourceServer.oidcLoginProviders,
+  oidcSendProviderApprovalEmail: resourceServer.oidcSendProviderApprovalEmail,
+  oidcProcessProviderApproved: resourceServer.oidcProcessProviderApproved,
   passIfNotAuthorized: resourceServer.passIfNotAuthorized,
   requirePrivilege: resourceServer.requirePrivilege,
   userToActor: resourceServer.userToActor,
@@ -42,8 +44,11 @@ module.exports = {
   checkImmer: oauthClient.checkImmer,
   checkImmerAndRedirect: oauthClient.checkImmerAndRedirect,
   handleOAuthReturn: [oauthClient.handleOAuthReturn, resourceServer.returnTo],
+  handleSamlReturn: [oauthClient.handleSamlReturn, resourceServer.returnTo],
   oidcPreRegister: oauthClient.oidcPreRegister,
-  oidcPostRegister: oauthClient.oidcPostRegister
+  oidcPostRegister: oauthClient.oidcPostRegister,
+  oidcPostMerge: oauthClient.oidcPostMerge,
+  samlProviderMetadata: oauthClient.samlServiceProviderMetadata
   // openId Connect server (WIP)
   /*
   oidcServerRouter: openIdServer.router,
