@@ -10,16 +10,20 @@ export function AvatarPreview ({ icon, avatar, size = 'large' }) {
     return null
   }
   const aviModelURL = ImmersClient.URLFromProperty(avatar?.url)
-  return (
-    <div className={c('avatarPreview', size)}>
-      <model-viewer
-        alt={avatar.name}
-        src={aviModelURL}
-        ar ar-modes='webxr scene-viewer quick-look'
-        poster={iconSrc}
-        shadow-intensity='1' camera-controls enable-pan
-      />
-      {avatar.name && <SanitizedHTML className='insetAvatar' html={avatar.name} />}
-    </div>
-  )
+  return aviModelURL
+    ? (
+      <div className={c('avatarPreview', size)}>
+        <model-viewer
+          alt={avatar.name}
+          src={aviModelURL}
+          ar ar-modes='webxr scene-viewer quick-look'
+          poster={iconSrc}
+          shadow-intensity='1' camera-controls enable-pan
+        />
+        {avatar.name && <SanitizedHTML className='insetAvatar' html={avatar.name} />}
+      </div>
+      )
+    : (
+      <img className={c('avatarPreview', size)} src={iconSrc} />
+      )
 }
